Organize ServerClientConsumer
